# server/openapi.yaml
openapi: 3.0.3
info:
  title: Cook Africa API
  description: |
    API backend de l'application de gestion interne du restaurant Cook Africa,
    plus le point d'entrée public utilisé par la page client `/commander`.

    ### Authentification
    La quasi-totalité des routes exigent un JWT Bearer obtenu via `POST /api/auth/login`,
    envoyé dans l'en-tête `Authorization: Bearer <token>`. Le token expire après 12h.

    Les routes sous `/api/public/*` sont volontairement **non authentifiées** : elles sont
    appelées depuis le navigateur d'un client final (page de commande en ligne), pas depuis
    un poste staff. Elles n'exposent que le menu disponible et la création/consultation de
    commandes ; toute autre donnée reste protégée.

    ### Rôles
    `admin`, `caissiere`, `serveur`, `cuisiniere`, `barman`, `caissier-en-ligne` — la plupart des routes
    restreignent l'accès à un sous-ensemble de rôles via un contrôle serveur (403 sinon).

    ### Restriction Wi-Fi
    Pour tout rôle autre que `admin`, les requêtes authentifiées sont en plus filtrées par
    IP si la restriction Wi-Fi est activée (`/api/wifi-config`) — 403 `wifi_restricted`
    si le poste n'est pas sur le réseau autorisé.

    ### Erreurs
    Toutes les erreurs suivent le même format `{ "error": "message" }`, avec le code HTTP
    approprié (400 validation, 401 non authentifié, 403 accès refusé, 404 introuvable,
    409 conflit, 429 limite de requêtes, 500 erreur serveur).
  version: 1.0.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      operationId: postApiAuthLogin
      tags: [Auth]
      summary: Connexion
      description: Limité à 10 tentatives / 15 min par IP. Bloqué si restriction Wi-Fi active et rôle ≠ admin.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, format: password }
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/UserPublic' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Identifiants valides mais poste hors réseau Wi-Fi autorisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: wifi_restricted }
                  message: { type: string }
...
  /api/auth/sessions:
    get:
      operationId: getApiAuthSessions
      tags: [Auth]
      summary: Journal d'audit des connexions (admin)
      parameters:
        - name: debut
          in: query
          schema: { type: string, format: date }
        - name: fin
          in: query
          schema: { type: string, format: date }
        - name: username
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Liste des sessions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Session' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/seed:
    post:
      operationId: postApiAuthSeed
      tags: [Auth]
      summary: Initialise les 5 comptes staff par défaut (une seule fois)
      description: |
        Protégé par un en-tête `X-Seed-Token` devant correspondre à la variable
        d'environnement `SEED_TOKEN` du serveur — pas par un JWT.
      security: []
      parameters:
        - name: X-Seed-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Utilisateurs créés
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  count: { type: integer, example: 5 }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Base déjà initialisée
  /api/auth/migrate-roles:
    post:
      operationId: postApiAuthMigrateRoles
      tags: [Auth]
      summary: Migre les anciens noms de rôles vers les nouveaux (admin)
      responses:
        '200':
          description: Migration effectuée
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/utilisateurs:
    get:
      operationId: getApiAuthUtilisateurs
      tags: [Auth]
      summary: Liste des utilisateurs staff (admin)
      responses:
        '200':
          description: Liste des utilisateurs (sans passwordHash)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: postApiAuthUtilisateurs
      tags: [Auth]
      summary: Créer un utilisateur staff (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Identifiant déjà utilisé
  /api/auth/utilisateurs/{id}:
    put:
      operationId: putApiAuthUtilisateursById
      tags: [Auth]
      summary: Modifier un utilisateur staff (admin)
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: Utilisateur mis à jour
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteApiAuthUtilisateursById
      tags: [Auth]
      summary: Désactiver un utilisateur staff (admin)
      description: Un admin ne peut pas désactiver son propre compte.
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Utilisateur désactivé
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/me:
    get:
      operationId: getApiAuthMe
      tags: [Auth]
      summary: Vérifie le token courant et retourne l'utilisateur
      responses:
        '200':
          description: Utilisateur courant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton obtenu via `POST /api/auth/login`, valable 12h.
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: Identifiant du document Firestore
      schema: { type: string }
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Non authentifié (token manquant, invalide ou expiré)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Accès refusé (rôle insuffisant, ou hors réseau Wi-Fi autorisé)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Role:
      type: string
      enum: [admin, caissiere, serveur, cuisiniere, barman, caissier-en-ligne]
...
    UserPublic:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        nom: { type: string }
        role: { $ref: '#/components/schemas/Role' }
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        prenom: { type: string }
        nom: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        actif: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLogin: { type: string, format: date-time, nullable: true }
    UserCreateRequest:
      type: object
      required: [nom, username, password, role]
      properties:
        prenom: { type: string }
        nom: { type: string }
        username: { type: string }
        password: { type: string, format: password }
        role: { $ref: '#/components/schemas/Role' }
    UserUpdateRequest:
      type: object
      properties:
        prenom: { type: string }
        nom: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        password: { type: string, format: password, description: Optionnel — change le mot de passe si fourni }
        actif: { type: boolean }
    Session:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        username: { type: string }
        nom: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        action: { type: string, enum: [login, logout] }
        ip: { type: string }
        timestamp: { type: string, format: date-time }
...