# server/routes/auth.js
const express = require('express');
const bcrypt = require('bcryptjs');
const jwt = require('jsonwebtoken');
const { db } = require('../firebase-admin');
const { authenticateToken, requireRole } = require('../middleware/auth');
const { pushNotification } = require('../utils/notifications');

const { isAllowedIp, getClientIp } = require('../utils/wifi');

const router = express.Router();

// Normalise les anciens noms de rôles vers les nouveaux
const ROLE_MIGRATION = {
  directeur:       'admin',
  receptionniste:  'caissiere',
  cuisinier:       'cuisiniere',
};

// caissier-en-ligne : accès restreint au dashboard + à l'onglet Commandes en
// Ligne (voir PAGE_ROLES côté client) — pour gérer les commandes du lien
// client à distance, sans passer par la facturation sur place.
const VALID_ROLES = ['admin', 'caissiere', 'serveur', 'cuisiniere', 'barman', 'caissier-en-ligne'];

async function logSession(userId, username, nom, role, action, ip) {
  await db.collection('sessions').add({
    userId, username, nom, role, action,
    ip: ip || 'inconnue',
    timestamp: new Date().toISOString()
  });
}

// POST /api/auth/login
router.post('/login', async (req, res) => {
  try {
    const { username, password } = req.body;
    if (!username || !password) {
      return res.status(400).json({ error: 'Identifiants requis' });
    }

    const snapshot = await db.collection('utilisateurs')
      .where('username', '==', username.toLowerCase().trim())
      .where('actif', '==', true)
      .limit(1).get();

    if (snapshot.empty) {
      return res.status(401).json({ error: 'Identifiants invalides' });
    }

    const userDoc = snapshot.docs[0];
    const user = userDoc.data();

    const passwordMatch = await bcrypt.compare(password, user.passwordHash);
    if (!passwordMatch) {
      return res.status(401).json({ error: 'Identifiants invalides' });
    }

    // Normaliser le rôle (migration anciens → nouveaux noms)
    const effectiveRole = ROLE_MIGRATION[user.role] || user.role;

    // Restriction Wi-Fi : l'admin peut se connecter depuis n'importe quel réseau
    if (effectiveRole !== 'admin' && !(await isAllowedIp(req))) {
      const clientIp = getClientIp(req);
      console.warn(`[WiFi] Tentative bloquée depuis ${clientIp} (X-Forwarded-For: ${req.headers['x-forwarded-for'] || 'absent'})`);
      return res.status(403).json({
        error: 'wifi_restricted',
        message: "Accès refusé : connectez-vous au réseau Wi-Fi de l'entreprise",
      });
    }

    await userDoc.ref.update({ lastLogin: new Date().toISOString() });
    await logSession(userDoc.id, user.username, user.nom, effectiveRole, 'login', req.ip).catch(() => {});

    pushNotification({
      type: 'info', icon: 'sign-in-alt',
      titre: 'Connexion utilisateur',
      message: `${user.nom} (${effectiveRole}) s'est connecté`,
      createdBy: user.username,
    });

    const nomComplet = user.prenom ? `${user.prenom} ${user.nom}`.trim() : user.nom;
    const token = jwt.sign(
      { id: userDoc.id, username: user.username, nom: nomComplet, role: effectiveRole },
      process.env.JWT_SECRET,
      { expiresIn: '12h' }
    );

    res.json({ token, user: { id: userDoc.id, username: user.username, nom: nomComplet, role: effectiveRole } });
  } catch (err) {
    console.error('Login error:', err);
    res.status(500).json({ error: 'Erreur serveur' });
  }
});

// POST /api/auth/logout
router.post('/logout', authenticateToken, async (req, res) => {
  try {
    await logSession(req.user.id, req.user.username, req.user.nom, req.user.role, 'logout', req.ip);
    res.json({ message: 'Déconnexion enregistrée' });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// GET /api/auth/sessions
router.get('/sessions', authenticateToken, requireRole('admin'), async (req, res) => {
  try {
    const { debut, fin, username } = req.query;
    const snap = await db.collection('sessions').orderBy('timestamp', 'desc').limit(500).get();
    let sessions = snap.docs.map(d => ({ id: d.id, ...d.data() }));
    if (debut)    sessions = sessions.filter(s => s.timestamp >= debut);
    if (fin)      sessions = sessions.filter(s => s.timestamp <= fin + 'T23:59:59');
    if (username) sessions = sessions.filter(s => s.username === username);
    res.json(sessions);
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// POST /api/auth/seed — initialisation des utilisateurs par défaut
// Protégé par un jeton d'amorçage (header X-Seed-Token) pour éviter qu'un tiers
// ne crée les comptes par défaut (mots de passe connus) avant le déploiement réel.
router.post('/seed', async (req, res) => {
  try {
    const seedToken = process.env.SEED_TOKEN;
    if (!seedToken || req.headers['x-seed-token'] !== seedToken) {
      return res.status(403).json({ error: 'Accès refusé' });
    }

    const snapshot = await db.collection('utilisateurs').limit(1).get();
    if (!snapshot.empty) {
      return res.status(409).json({ error: 'Base déjà initialisée' });
    }

    const utilisateurs = [
      { username: 'admin',      nom: 'Admin Système',   role: 'admin',      password: 'Admincookaf@2026!' },
      { username: 'caissiere',  nom: 'Caissière',       role: 'caissiere',  password: 'Caissecookaf@2026!' },
      { username: 'serveur',    nom: 'Serveur',          role: 'serveur',    password: 'Servcookaf@2026!' },
      { username: 'cuisiniere', nom: 'Cuisinière',       role: 'cuisiniere', password: 'Cuisincookaf@2026!' },
      { username: 'barman',     nom: 'Barman Service',   role: 'barman',     password: 'Barmancookaf@2026!' },
    ];

    const batch = db.batch();
    for (const u of utilisateurs) {
      const ref = db.collection('utilisateurs').doc();
      const passwordHash = await bcrypt.hash(u.password, 10);
      batch.set(ref, {
        username: u.username,
        nom: u.nom,
        role: u.role,
        passwordHash,
        actif: true,
        createdAt: new Date().toISOString(),
        lastLogin: null
      });
    }
    await batch.commit();

    res.json({ message: 'Utilisateurs créés', count: utilisateurs.length });
  } catch (err) {
    console.error('Seed error:', err);
    res.status(500).json({ error: 'Erreur serveur' });
  }
});

// POST /api/auth/migrate-roles — migration des anciens rôles vers les nouveaux
router.post('/migrate-roles', authenticateToken, requireRole('admin'), async (req, res) => {
  try {
    const snap = await db.collection('utilisateurs').get();
    const batch = db.batch();
    let count = 0;
    snap.docs.forEach(doc => {
      const role = doc.data().role;
      const newRole = ROLE_MIGRATION[role];
      if (newRole) {
        batch.update(doc.ref, { role: newRole, updatedAt: new Date().toISOString() });
        count++;
      }
    });
    if (count > 0) await batch.commit();
    res.json({ message: `${count} utilisateur(s) migré(s)` });
  } catch (err) {
    console.error('Migrate roles error:', err);
    res.status(500).json({ error: 'Erreur serveur' });
  }
});

// GET /api/auth/utilisateurs — liste des utilisateurs (admin)
router.get('/utilisateurs', authenticateToken, requireRole('admin'), async (req, res) => {
  try {
    const snap = await db.collection('utilisateurs').orderBy('createdAt', 'desc').get();
    const users = snap.docs.map(d => {
      const { passwordHash, ...rest } = d.data();
      return { id: d.id, ...rest };
    });
    res.json(users);
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// POST /api/auth/utilisateurs — créer un utilisateur (admin)
router.post('/utilisateurs', authenticateToken, requireRole('admin'), async (req, res) => {
  try {
    const { prenom, nom, username, password, role } = req.body;
    const validRoles = VALID_ROLES;
    if (!nom || !username || !password || !role) {
      return res.status(400).json({ error: 'Nom, identifiant, mot de passe et rôle requis' });
    }
    if (!validRoles.includes(role)) {
      return res.status(400).json({ error: 'Rôle invalide' });
    }
    const existing = await db.collection('utilisateurs')
      .where('username', '==', username.toLowerCase().trim()).limit(1).get();
    if (!existing.empty) {
      return res.status(409).json({ error: 'Cet identifiant est déjà utilisé' });
    }
    const passwordHash = await bcrypt.hash(password, 10);
    const data = {
      prenom: prenom?.trim() || '',
      nom: nom.trim(),
      username: username.toLowerCase().trim(),
      role,
      passwordHash,
      actif: true,
      createdAt: new Date().toISOString(),
      lastLogin: null,
    };
    const ref = await db.collection('utilisateurs').add(data);
    const { passwordHash: _, ...safe } = data;
    res.status(201).json({ id: ref.id, ...safe });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// PUT /api/auth/utilisateurs/:id — modifier un utilisateur (admin)
router.put('/utilisateurs/:id', authenticateToken, requireRole('admin'), async (req, res) => {
  try {
    const { prenom, nom, role, password, actif } = req.body;
    const validRoles = VALID_ROLES;
    const update = { updatedAt: new Date().toISOString() };
    if (nom !== undefined)   update.nom = nom.trim();
    if (prenom !== undefined) update.prenom = prenom.trim();
    if (role !== undefined) {
      if (!validRoles.includes(role)) return res.status(400).json({ error: 'Rôle invalide' });
      update.role = role;
    }
    if (actif !== undefined) update.actif = actif;
    if (password) update.passwordHash = await bcrypt.hash(password, 10);
    await db.collection('utilisateurs').doc(req.params.id).update(update);
    res.json({ message: 'Utilisateur mis à jour' });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// DELETE /api/auth/utilisateurs/:id — désactiver un utilisateur (admin)
router.delete('/utilisateurs/:id', authenticateToken, requireRole('admin'), async (req, res) => {
  try {
    if (req.params.id === req.user.id) {
      return res.status(400).json({ error: 'Vous ne pouvez pas désactiver votre propre compte' });
    }
    await db.collection('utilisateurs').doc(req.params.id).update({
      actif: false,
      updatedAt: new Date().toISOString(),
    });
    res.json({ message: 'Utilisateur désactivé' });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

// GET /api/auth/me — vérifie si le token est valide et retourne l'utilisateur courant.
// Utilisé par le client pour valider un token sauvegardé avant de restaurer la session.
router.get('/me', authenticateToken, (req, res) => {
  res.json({ id: req.user.id, username: req.user.username, nom: req.user.nom, role: req.user.role });
});

// GET /api/auth/check-wifi — vérifie si le client est sur le réseau Wi-Fi autorisé.
// Utilisé par le client pour la vérification proactive toutes les 30 s.
// Pour les non-admin, authenticateToken a déjà rejeté la requête si hors Wi-Fi.
router.get('/check-wifi', authenticateToken, (_req, res) => {
  res.json({ ok: true });
});

module.exports = router;